feat: async by default, safer port updates
