+++ Cargo.toml
@@ -4,25 +4,20 @@ version = "0.3.0"
 edition = "2021"
 license = "Apache-2.0 OR MIT"
 repository = "https://github.com/kiranshila/tftp_client"
-description = "A pure-rust TFTP client library"
+description = "A pure-rust async TFTP client library"
 homepage = "https://github.com/kiranshila/tftp_client"
 readme = "README.md"
 keywords = ["tftp"]
 categories = ["network-programming"]
 rust-version = "1.78"
 
-[features]
-default = []
-async = ["dep:async-net", "dep:async-io", "dep:futures-lite"]
-
 [dependencies]
 byte-strings = "0.3"
 thiserror = "2"
 tracing = "0.1"
-
-async-net = { version = "2.0", optional = true }
-async-io = { version = "2.4", optional = true }
-futures-lite = { version = "2.6", optional = true }
+async-net = "2"
+async-io = "2"
+futures-lite = "2"
 
 [dev-dependencies]
 paste = "1"
